requirements.txt
affine==2.3.0
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
docopt==0.6.2
Fiona==1.8.20
GDAL==3.3.2
GeoAlchemy2==0.9.4
geopandas==0.9.0
geoserver-rest==1.2.5
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
Js2Py==0.71
kiwisolver==1.3.2
matplotlib==3.4.3
munch==2.5.0
numpy==1.21.2
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pipwin==0.5.1
postgres-helper==0.0.3
psycopg2==2.9.1
pycurl==7.43.0.5
Pygments==2.10.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pyproj==3.2.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
rasterio==1.2.8
requests==2.26.0
Rtree==0.9.7
scipy==1.7.1
seaborn==0.11.2
Shapely==1.7.1
six==1.16.0
snuggs==1.4.7
soupsieve==2.2.1
SQLAlchemy==1.4.23
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.5.0
